+++ src/main.rs
@@ -6,9 +6,11 @@ fn main() {
     let mut max = 7.5;
     let nr_runs = 150;
     let space = 60;
+    let unit = String::from("");
 
     for _ in 0..nr_runs {
         max *= 1.05;
-        println!("{}", float_axis_labels(min, max, space, 1));
+        let labels = float_axis_labels(min, max, space, 1, &unit);
+        println!("{}", labels);
     }
 }